+++ src/polydoc/__init__.py
@@ -83,7 +83,7 @@ from .model import (
     plain,
 )
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 
 __all__ = [
     # API